Initial geometry restructure (#2063)
* I moved calculate_distances to a new folder, packet_transfer. Dependencies were renamed.

* I moved calculate_distances to a new folder, transport. Dependencies were renamed.

* I added myself to the .mailmap
