+++ .mailmap
@@ -141,6 +141,8 @@ Nathan Kowalski <56104113+kowal180@users.noreply.github.com>
 
 Nilesh Patra <37436956+nileshpatra@users.noreply.github.com>
 
+Nolan Brown <nolanbrown01@gmail.com>
+
 Pratik Patel <pratikpatel15133@gmail.com>
 Pratik Patel <pratikpatel15133@gmail.com> Pratik151 <pratikpatel15133@gmail.com>
