+++ tardis/montecarlo/montecarlo_numba/vpacket.py
@@ -17,7 +17,7 @@ from tardis.montecarlo.montecarlo_numba.r_packet_transport import (
     move_packet_across_shell_boundary,
 )
 
-from tardis.montecarlo.montecarlo_numba.calculate_distances import (
+from tardis.transport.calculate_distances import (
     calculate_distance_boundary,
     calculate_distance_line,
 )
